requirements.txt
asgiref==3.3.1
Django==3.1.6
django-multiselectfield==0.1.12
djangorestframework==3.12.2
joblib==1.0.1
numpy==1.20.1
pandas==1.2.2
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.1
six==1.15.0
sqlparse==0.4.1
threadpoolctl==2.1.0
